Implemented bunch of instruction/attributes, fixed bug with byte reading

+++ src/main/java/org/bytecodeparser/constant/Tag.java
@@ -8,5 +8,5 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.TYPE)
 public @interface Tag {
-    byte value();
+    short value();
 }
